requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.35.0
scipy>=1.11.0
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0
gradio>=4.0.0
accelerate>=0.24.0
gtts>=2.5.0
pydub>=0.25.0
yt-dlp>=2023.0.0
requests>=2.31.0

# YouTube API (opsiyonel - sadece video yükleme için)
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

